requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
dj-rest-auth==7.2.0
Django==6.0.4
django-allauth==65.16.1
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.12
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pydivert==3.1.1
PyJWT==2.12.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
